added correct sizeof(ttl) in setsockopt(...)

+++ src/traceroute.c
@@ -88,7 +88,7 @@ int traceroute(char* dest_hostname){
        freeaddrinfo(dest_addrinfo_collection);
        return EXIT_FAILURE;
      }
-    error = setsockopt(send_socket, IPPROTO_IP, IP_TTL, &ttl, sizeof(int));
+    error = setsockopt(send_socket, IPPROTO_IP, IP_TTL, &ttl, sizeof(unsigned short));
     if(error != 0){
        perror("\nerror setting socket options");
        freeaddrinfo(dest_addrinfo_collection);
